+++ exercises/week 5/solutions/bit-aloo/chapter-8/script.py
