+++ exercises/week 5/solutions/bit-aloo/chapter-9/helper.py
@@ -127,7 +127,11 @@ def bits_to_target(bits):
     '''Turns bits into a target (large 256-bit integer)'''
     # last byte is exponent
     exponent = bits[-1]
+    # the first three bytes are the coefficient in little endian
+    coefficient = little_endian_to_int(bits[::2])
     coefficient = little_endian_to_int(bits[:-1])
+    # the formula is:
+    # coefficient * 256**(exponent-3)
     return coefficient * 256**(exponent - 3)
 
 
@@ -151,16 +155,25 @@ def calculate_new_bits(previous_bits, time_differential):
     '''Calculates the new bits given
     a 2016-block time differential and the previous bits'''
     # if the time differential is greater than 8 weeks, set to 8 weeks
-    # if the time differential is less than half a week, set to half a week
-    # the new target is the previous target * time differential / two weeks
-    # if the new target is bigger than MAX_TARGET, set to MAX_TARGET
-    # convert the new target to bits
     if time_differential > TWO_WEEKS * 4:
         time_differential = TWO_WEEKS * 4
-    if time_differential < TWO_WEEKS // 4:
-        time_differential = TWO_WEEKS // 4
+    
+    # if the time differential is less than half a week, set to half a week
+    if time_differential < TWO_WEEKS//4:
+        time_differential = TWO_WEEKS//4
+    
+    # the new target is the previous target * time differential / two weeks
     new_target = bits_to_target(previous_bits) * time_differential // TWO_WEEKS
+    # convert the new target to bits
     return target_to_bits(new_target)
+
+"""
+ Comment, there is no MAX_TARGET variable definite, seems to be an issue with this exercise
+    # if the new target is bigger than MAX_TARGET, set to MAX_TARGET
+    if new_target > MAX_TARGET:
+        new_target = MAX_TARGET
+"""
+    
 
 
 class HelperTest(TestCase):
