+++ exercises/week 5/solutions/bit-aloo/chapter-9/op.py
